refactor: Move meta image types
