--- design-system/components/meta-image/meta-image.types.tsx
@@ -1,8 +1,0 @@
-import { Theme } from '@/types'
-
-export type MetaImageProps = {
-  title: string
-  subtitle: string
-  theme: Theme
-  backgroundURL?: string
-}
